Move book edit page component (#178)
