+++ frontend/app/routes/home.books.$bookId.edit/components/BookDetailEditTitle.tsx
