+++ frontend/app/routes/home.books.$bookId.edit/components/BookDetailIsbnForm.tsx
