+++ frontend/app/routes/home.books.$bookId.edit/components/BookDetailPublishedDateForm.tsx
