+++ frontend/app/routes/home.books.$bookId.edit/components/BookDetailStockForm.tsx
